refactor(compiler-plugin): move `StackFunctionCall` to the backend
